+++ packages/studio/src/app-next/writing-workbench/index.ts
@@ -3,4 +3,5 @@ export * from "./WorkbenchResourceTree";
 export * from "./WorkbenchTopBar";
 export * from "./WorkbenchWritingActions";
 export * from "./WritingWorkbenchRoute";
+export * from "./resource-viewers";
 export * from "./useWorkbenchResources";